Markup: Ensure small markups are deleted on server
Some weird hackery in here. Basically, we want to destroy
this.cur_markup, but we've probably got a create() request to server
already in flight. If we destroy it *before* that create() finishes,
we'll destroy it client side without destroying it server side. (We
won't don't yet know the id to destroy it on the server.) That means
it'll still be created on the server and it'll show up on page reloads.

Solution? Don't destroy until our create() is done