Adding conceptual code for watermark
The code is using an overlay image then drawing the text on it, then
plastering that layer on top of the original image, half way down and on
the far right.  Currently I'm doing like 39% width and 35% height...
It's all quite customizable, unlike apple products.  It also has a
plethora of code in it that needs to be changed, but I just wanted to
see it done.  Secondly I stole some of the layer merging from random
places on the internet because I didn't know what PIL was.  Thirdly it
appears to not work with the version of python installed for Mr.
Picdoc...