Finishing touches for general instructions
Implemented the handler, screwed around with the model, learned how to
use pdb.

Additionally I changed some of the css, mostly areas that said, ****DO
NOT TOUCH OR YOU WILL BREAK EVERYTHING***.  But in all seriousness I
shifted the instructions over to the left, they were starting from the
left of the picture and going out 700 px to the right.  I shifted the
desc and desc_titles over to the left...