Fixing focus yet again
I really feel guilty about how it's fixed, essentially I stop the
re-rendering of the markup and text-box.  Easier said than done, since
it has to keep rendering while you are dragging it.  I could be stupid
and shove something into the model such as 'render', True and then when
it's finished creating the markup change that to False, but then I'd be
considered breaking the mvc of the client side code.  Bad Llama