Implemented csrf for upload and markup pages
Despite their recommendation of using the cookie (which I read after the
fact) I dislike relying on cookies being enabled.  I realize the site so
far is reliant on cookies, but it's just as easy to push the csrf_token
into the base page and not worry about whether or not it's available in
the dom for use in setting your headers.  When using Backbone you need
to override the sync method and set the header before sending it, when
using jquery ajax call you can just set the header in the call.  In the
form you just need to include the {% csrf_token %} somewhere in the form
and it will be included in the post meaning you can enable csrf_protect
on your handler.