Adding Group Model, Remake Group Fix
The Grouping is now only remade when they've changed the grouping and
there are no grouping records in the database.  Beside that I think I'd
like to add a fk from the pic to the group, but it's being a db right
now and complaining about circular dependencies and how it's mother is a
whore.