Removes batch id from the markup url
I'm not sure why we had the batch id in the url in the first place.
In any case, leaving the exposed batch_id in the url is ugly, and allows
people to look at other people's batches. I suppose we could verify that
the url batch id mathes the real batch id, but if we are going to do
that, why not just use the real batch id anyway?