Adds merge batch page (which is still empty)
Batches may require merging if the user signs in and out at the right
time. We need to merge. This is the initial commit to get that done, but
the page is still empty.

Also, unrelated, I reordered some of the stuff in common/models.py
because it was bugging me. All imports are at the top, and I also
alphabetized the subsections of imports because vim can do that in 1
second and it's fun.