Small modifications to Job Statuses
Changed user_rejects to user_rejected, and user_requests_additional_work
to user_requests_modification.  Also added code to remove a job row.

I'd like to also add the actions to a model and link them all in using
backbone etc.  That way I could remove an action or change a status of a
row without having to refresh the page.