Adds get_time_string to common.functions
Time zones are a pain, so we should use get_time_string to format times
that we'll print to the user. It return a string for the given time a la
facebook -- "2 hours ago", "on 10 Oct 2012", or "5 minutes ago".

This commit also contains a few more little things for the merging
(which still doesn't work) and a few scattered cleanup things.