Markup Aesthetic changes
No more mini markups when deleting other markups

Fun popover when we're in readonly mode over the markups, so you don't
have to look down below to see the instructions

When you create a new markup the focus is automatically put into that
markups description (this makes more sense when you are a real user,
creating real markups).  If you are just screwing around and creating
random markups it's partially annoying (probably for the likes of you
and me ;) ).  I'm fine if we want to remove this... Just adding extra
fun stuff.