Approval by Moderator mostly functional
If you add yourself to the 'Album Moderators' group then you will see a
link at the top of the website.  When you click on this link you see all
the albums that aren't approved that need approving.

I also am generating the permissions and 'Album Moderators' group
automagically in the init of the common