Fix popover and minor null fixes
While reading the stupid twitter thing I noticed the default trigger for
popover was click.  Set it to hover, and kablamo, all is well..

At the same time I was writing out urls to get to a page to try
something and some of my logic was crashing (not the worst thing in the
world, crash is better than them randomly getting access), but I like to
fail silently when I'm aware of the issue.  Reordered and all honkey
dorey now