Revamp of Messaging & Job Workflow
There are still quite a few quirks I'd like to workout when it comes to
messaging and the Job Workflow, let's list the ones that come to behind
before I go to sleep:

When either user writes a message we flip some sort of flag, I'm still
working on wether or not it will change the job state.  Issue with it
changing the job state is we almost need to know what state it was in
before, so they can return to said state, or even have the same actions
available in the pre-state.  So since I don't like that, it would be
some sort of flag flipping when someone comments.  What happens when
they've commented after the doctor's submitted their work?  Would this
automatically set the state to user wants additional work done?  If it's
on a specific picture I'd say yes... so ya know, figure that out future
Daniel.

Job Workflow:  Most of the issues surrounding this are related to the
messaging as well.

Message Page:  This doesn't exist anymore!  Ever since messaging turned
into contact and it's related to each picture... now that I contemplate
that, I'm not sure if contact is the right wording.  Contact Concerning
Output.  Contact Concerning Image.  Something like that, instead of just
Contact Doctor.  Contact Doctor/User will be reserved for just generic,
hey, you are cool, don't talk to me like that, screw you!!!!

Okay, I can't think anymore, off to bed