Actually place a hold on cards
We've added wrapper classes for some of the Balanced Payments classes to
access their objects. We locally store some things since the fetch is so
slow.

This commit also gets rid of the Charge object, which was associated
with stripe.

As a side note, I removed a "null=True" on a CharField. We shouldn't
ever have that on a CharField because it leaves gets confusing if blank
is "" or null, which aren't equal in db land. Django docs talk about
this. Don't let me do it again.