Quick spot check testing/changes
I changed it to when they click login they have a login prompt instead
of a create prompt.  Now if they are pushed to the login page by another
page, they are prompted with a create by default.

Fixed an issue with the reject views in where you could never make it to
the view since I was missing the only positive case... Not sure how I
missed it, I'll blaim it on Riley