Adds account settings page
The structure of the account settings page is a little confusing.
Basically, both users and doctors go to '/account_settings/' but they
see different things when they get there. So there's a common page for
both and individual content for either.

The pages are still needing some love. As of this commit, password reset
and email change don't work, and I'm still playing with the pages for
the doctor's bank account. Give me a minute.