Credit doctors bank accounts
Adds withdraw page logic, which requires the user to request a bank
transfer. We don't allow withdrawals below $5, which hopefully will
minimize the impact of ACH fees. Note: We currently pay this fee. I want
to keep it easy on the doctors, plus it simplifies the payment
processing logic by quite a bit.