Load settings based on [deploy_type].cfg file
Want to deploy a production server? touch settings/production.cfg
Want to deploy a test server? touch settings/test.cfg
Similarly for sandbox. dev is still the default.

Note that *.cfg files are in the .gitignore, so this is a setting that's
totally local to your environment. I don't want uncommitted changes
scaring me on the server.

We actually were already doing this through uwsgi (see the setting in
uwsgi_picdoctorsapp.ini), but it wasn't set for commandline stuff, such
as 'python manage.py syncdb', so I decided it should be fixed
everywhere.