Add restoration example to index.html
This commit also lines up botanical before/after pictures.  I like it
fading from before to after without changing the orientation or cropping
at all. It makes for a pretty cool fade in effect.

This also makes the hero-unit separate from the transitions so it can
always be clicked, but the transitions still happened. Previously we
reset our counter every time and could click during transitions, so it
was really ugly. This is much better.