Fix stupid redirects... finally...
I spent way to long trying to figure this out. In the end, I just read
the django sources for where the logic is used and figured out that we
could use a header that wasn't getting stripped.

I think the primary problem is that nginx and uwsgi don't communicate
over http -- they communicate over uwsgi, which is some weird binary
protocol that's way faster. But it doesn't explain why blog posts on the
interwebs think the X-Forwarded-Protocol thing should work. I'm not sure
I totally get it, and I'm okay with that.