Don't allow deletion until picture is fully uploaded
We no longer allow a user to delete a picture while it is uploading.

This is good because: 1) I don't have to dive into the evil hackery that
is the upload page to cancel an upload. 2) I don't have to deal with
weird state issues if the picture is uploaded, but the resizing is in
progress 3) I don't have a decent way to a server side delete without the
uuid, which I don't have yet anyway.

This is bad because you can't cancel an upload until it's done.

I'm willing to live with this for now.