Add emailer app, debug_spam_emails debug page
So, my wife wants to get all creative like with emails. At least, she
does once she's not permanantly miserable because she's going to have a
baby any day. Anyway, she wants to tweak the emails a lot, and that's
way harder right now than it should be. This commit:
  - Consolidates the widely copy/pasted email code into one sane app
  - Adds a basic (admin only) page for sending spam emails