Embed subject, plaintext, and html within an email template
As part of my cleanup effort, I really want the emails to keep all the
friggin email info in the email template, and I had to use some weird
tricks to get it to work. This is a much cleaner solution.

I still need to move all the emails (besides reset_password_email.html)
to use this approach.