Add logging to file, better logging
Log to file and the console. It's pretty. log.debug messages do NOT go
to the console, so as to not spam it too much.

Want logging? Put this at the top of your file:

import logging; log = logging.getLogger('pd')

What? Can't remember that? Well, that's what <control+x><control+l> is
for, doofus. What are we using vim for, anyway? (Hopefully it's already
at the top of some other buffer that you had open...)

Then just do:

log.debug(), log.info(), log.error(), etc

Notes: This logs to 'logfile' in the home directory of pd. I'm planning
on moving it to a more persistant place in /var. That's a small change,
but I can't do it now.