Fix server side log
Sorry to pollute the history with these, but I'm testing if this works
with deploys on the production server, so the commit has to be upstream.

Note that this also adds *.log to the .gitignore, so you can do
something like this:

ln -s /var/log/django/picdoctors.log /code/picdoctors/pd.log

That way you can see the log file without hunting around you filesystem
for it.