Fix django-ember-precompile path
A while back I decided that all the node.js stuff should be global so I
didn't have to be annoyed with a local 'node_modules' folder full of
binaries that weren't really part of picdoctors. I still think that was
a good decision, but if you disagree I'm fine with that.

But, anyway, I moved django-ember-precompile to the global path, like
the other node modules (use 'sudo' and add '-g' to the install -- 'sudo
npm install -g django-ember-precompile'), and I didn't change the path
in the settings, so this fixes that.