Handle bad file uploads more gracefully
If there's any problem with an upload (pdf, etc) we handle it gracefully
on the server and display a nice error on the page for the user.
Unfortunately, our attempt to create a valid <canvas> from a pdf makes
pic_container go wonky during the upload, but I'll deal with that later.