Don't promise non US countries
Change field to not promise non US countries yet. To quote the stripe
docs, referring to the transers API:

"This is currently only available for US Stripe accounts sending funds
to third-party US bank accounts or US Visa or MasterCard debit cards
that aren't prepaid."

We can do transfers to other countries, via Stripe Connect, but that
won't be in place by the 6th. Also, they've got beta support for
bitcoin, but it's not there yet.