Fix fees calculations for doctors
In the massive refactor, I wanted to make sure the stripe transaction
went through before I updated the job accordingly. However, in doing so,
I wasn't calculating the doctor's payout price until after I made the
strip transaction, so the doctor's payout price was still 0.

Fixed now. And I've confirmed that we only have to pass the fees when
creating the uncaptured charge -- we don't have to apply fees again when
capturing the charge.