added some ui to make stuff even more obvious
Feel free to change some of these, I moved some of the banners on the
pay page cause they kind of were in the way (imho).

I also changed some of the wording to setting an offer instead of
paying, it seems less intimidating and is honestly closer to what
they're doing.