Verify card works during set price
Until now, we don't actually verify that their card will work until the
doctor tries to accept the job. Once we (then) find out that their card
is invalid, it creates a bit of a pain for them to pay for the job
correctly.

To solve this, we now create a hold (uncaptured charge) for $1 and
immediately release it when they create the job. This should help with
declined cards.