Add cc holds info to FAQ, remove from set price page
Now that we only do a $0.50 authorization hold, I'd rather not scare
anyone away from the set price page with scary language. Even though
that's less informative, I hope you're okay with it.

I did add a lot more info to the FAQ, which I'll point them to if they
are upset about the $0.50 hold.